+++ src/test/java/com/ivanbasic/learnspring/controller/webmvctest/HomeControllerTest.java
@@ -1,0 +1,57 @@
+package com.ivanbasic.learnspring.controller.webmvctest;
+
+
+import com.ivanbasic.learnspring.configuration.SecurityConfig;
+import com.ivanbasic.learnspring.controller.AuthController;
+import com.ivanbasic.learnspring.controller.HomeController;
+import com.ivanbasic.learnspring.service.TokenService;
+
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+
+@WebMvcTest({HomeController.class, AuthController.class})
+@Import({SecurityConfig.class, TokenService.class})
+class HomeControllerTest {
+
+    @Autowired
+    MockMvc mvc;
+
+    @Test
+    void rootWhenUnauthenticatedThen401() throws Exception {
+        this.mvc.perform(get("/"))
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    void rootWhenAuthenticatedThenSaysHelloUser() throws Exception {
+        MvcResult result = this.mvc.perform(post("/token")
+                        .with(httpBasic("ivan", "i")))
+                .andExpect(status().isOk())
+                .andReturn();
+
+        String token = result.getResponse().getContentAsString();
+
+        this.mvc.perform(get("/")
+                        .header("Authorization", "Bearer " + token))
+                .andExpect(content().string("Hello, ivan"));
+    }
+
+    @Test
+    @WithMockUser
+    public void rootWithMockUserStatusIsOK() throws Exception {
+        this.mvc.perform(get("/")).andExpect(status().isOk());
+    }
+
+}